PR for bump node 14 branch into master

+++ db/models/chunk.js
@@ -1,9 +1,12 @@
-const crypto = require('crypto');
 const Model = require('../model');
 const fs = require('fs');
 const path = require('path');
+const Sequelize = require('sequelize');
+const _ = require('lodash');
+const utils = require('#utils');
 let StorageLink;
 let File;
+let FileMap;
 
 class Chunk extends Model {
     constructor(...args) {
@@ -12,19 +15,8 @@ class Chunk extends Model {
         // This is to avoid circular dependencies:
         File = require('./file');
         StorageLink = require('./storage_link');
+        FileMap = require('./file_map');
     }
-
-    static _buildIndices() {
-        this._addIndex('ul_status');
-        this._addIndex('dl_status');
-    }
-
-    /**
-     * Fields:
-     * - id which is chunk content's hash
-     * ----
-     * - known_providers
-     */
 
     getData() {
         // todo: read from fs if you have it already or retrieve using storage layer client
@@ -32,7 +24,8 @@ class Chunk extends Model {
     }
 
     setData(rawData) {
-        let hash = this.ctx.utils.hashFnHex(rawData);
+        if (!Buffer.isBuffer(rawData)) throw Error('Chunk.setData: rawData must be a Buffer');
+        let hash = utils.hashFnHex(rawData);
 
         if (this.id) {
             if (this.id !== hash) {
@@ -45,28 +38,26 @@ class Chunk extends Model {
         }
 
         this.dl_status = Chunk.DOWNLOADING_STATUS_DOWNLOADED;
-        this.length = rawData.length;
+        this.size = rawData.size;
 
         Chunk.forceSaveToDisk(rawData, this.id);
     }
 
-    getLength() {
-        if (typeof this.length !== 'undefined') {
-            return this.length;
+    getSize() {
+        if (typeof this.size !== 'undefined' && this.size !== null) {
+            return this.size;
         } else {
             const filePath = Chunk.getChunkStoragePath(this.id);
-            if (!fs.existsSync(filePath)) return undefined;
+            if (!fs.existsSync(filePath)) throw new Error('cannot stat file '+filePath); // todo: sanitize
             return fs.statSync(filePath).size;
         }
     }
 
     static async findOrCreateByData(rawData) {
-        let id = this.ctx.utils.hashFnHex(rawData);
-        let result = await this.find(id);
-        if (result === null) {
-            result = this.new();
-            this.id = id;
-        }
+        if (!Buffer.isBuffer(rawData)) throw Error('Chunk.findOrCreateByData: rawData must be a Buffer');
+
+        let id = utils.hashFnHex(rawData);
+        let result = await this.findOrCreate(id);
 
         result.setData(rawData);
 
@@ -82,7 +73,7 @@ class Chunk extends Model {
 
     async reconsiderUploadingStatus(cascade = true) {
         let fn = async() => {
-            const live_copies = await this.storage_links.signed;
+            const live_copies = await StorageLink.byChunkIdAndStatus(this.id, StorageLink.STATUS_SIGNED);
 
             // 1. Redundancy
             if (live_copies.length < this.redundancy) {
@@ -117,6 +108,10 @@ class Chunk extends Model {
         // todo: maybe immediately start uploading here? trigger the tick()? but be careful with recursion loops
     }
 
+    async getLinksWithStatus(status) {
+        return await StorageLink.byChunkIdAndStatus(this.id, status);
+    }
+
     async reconsiderDownloadingStatus(cascade = true) {
         // todo todo todo
 
@@ -135,10 +130,18 @@ class Chunk extends Model {
     }
 
     async getFiles() {
-        if (! ('belongsToFiles' in this._attributes)) return [];
-        const file_ids = this.belongsToFiles.map(x => x[0]);
-        let results = await Promise.all(file_ids.map(async(id) => await File.find(id)));
-        return results;
+        const filemaps = await FileMap.findAll({
+            attributes: ['id', 'file_id'],
+            where: { 'chunk_id': this.id }
+        });
+
+        const fileIds = _.uniq(filemaps.map((element) => element.file_id));
+
+        let files = [];
+        for(let fileId of fileIds) {
+            files.push(await File.findOrFail(fileId));
+        }
+        return files;
     }
 
     async changeULStatus(status) {
@@ -152,17 +155,13 @@ class Chunk extends Model {
         await this.save();
     }
 
-    addBelongsToFile(file, offset) {
-        if (! ('belongsToFiles' in this._attributes)) {
-            this._attributes.belongsToFiles = [];
-        }
-
-        // Remove duplicates
-        for (let b of this._attributes.belongsToFiles) {
-            if (b[0] === file.id && b[1] === offset) return;
-        }
-
-        this._attributes.belongsToFiles.push([ file.id, offset, file.originalPath ]);
+    // This fn doesn't save!
+    async addBelongsToFile(file, offset) {
+        const filemap = FileMap.build();
+        filemap.file_id = file.id; // todo: make sure file.id is not empty
+        filemap.chunk_id = this.id;
+        filemap.offset = offset;
+        await filemap.save();
     }
 
     getStorageLinks() {
@@ -174,7 +173,7 @@ class Chunk extends Model {
                 if (! allStatuses.includes(status) && status !== 'all') return void 0;
                 if (status === 'all') {
                     let results = await Promise.all(allStatuses.map(async(s) => await this.getStorageLinks()[s]));
-                    return this.ctx.utils.iterableFlat(results);
+                    return utils.iterableFlat(results);
                 }
                 return await StorageLink.byChunkIdAndStatus(this.id, status);
             },
@@ -197,16 +196,20 @@ class Chunk extends Model {
 
     static getChunkStoragePath(id) {
         const cache_dir = path.join(this.ctx.datadir, this.ctx.config.client.storage.cache_path);
-        this.ctx.utils.makeSurePathExists(cache_dir);
+        utils.makeSurePathExists(cache_dir);
         return path.join(cache_dir, 'chunk_' + id);
     }
 
     static forceSaveToDisk(data, id = null) {
-        if (id === null) id = this.ctx.utils.hashFnHex(data);
+        if (!Buffer.isBuffer(data)) throw Error('Chunk.forceSaveToDisk: data must be a Buffer');
+
+        if (id === null) id = utils.hashFnHex(data);
 
         // todo: dont zero out the rest of the chunk if it's the last one, save space
 
         // todo: what if already exists? should we overwrite again or just use it? without integrity check?
+        // todo: important question! because there was a bug related to that, the file was saved under the wrong id and until we cleared the file system we didn't know how to debug it
+        // todo: but if we decide to rewrite here and something goes wrong, we might accidentally rewrite the good version. important for providers not to destroy information, just in case
         const chunk_file_path = Chunk.getChunkStoragePath(id);
         if (! fs.existsSync(chunk_file_path)) {
             fs.writeFileSync(chunk_file_path, data, { encoding: null });
@@ -215,9 +218,6 @@ class Chunk extends Model {
         return id;
     }
 
-    getStatus() {
-        throw Error('Dev message: Which status? ul_status or dl_status');
-    }
 }
 
 Chunk.UPLOADING_STATUS_CREATED = 'us0';
@@ -227,5 +227,25 @@ Chunk.DOWNLOADING_STATUS_CREATED = 'ds0';
 Chunk.DOWNLOADING_STATUS_DOWNLOADING = 'ds1';
 Chunk.DOWNLOADING_STATUS_DOWNLOADED = 'ds99';
 Chunk.DOWNLOADING_STATUS_FAILED = 'ds2';
+
+Chunk.init({
+    id: { type: Sequelize.DataTypes.STRING, unique: true, primaryKey: true },
+    size: { type: Sequelize.DataTypes.INTEGER, allowNull: true },
+
+    ul_status: { type: Sequelize.DataTypes.STRING, defaultValue: Chunk.UPLOADING_STATUS_CREATED },
+    dl_status: { type: Sequelize.DataTypes.STRING, defaultValue: Chunk.DOWNLOADING_STATUS_CREATED },
+
+    // allowNull in the next rows because this is only related to uploading and storing, not chunks queued for downloading
+    redundancy: { type: Sequelize.DataTypes.INTEGER, allowNull: true },
+    expires: { type: Sequelize.DataTypes.BIGINT, allowNull: true },
+    autorenew: { type: Sequelize.DataTypes.BOOLEAN, allowNull: true },
+}, {
+    indexes: [
+        { fields: ['ul_status'] },
+        { fields: ['dl_status'] },
+    ]
+});
+
+// Chunk.belongsTo(File, { foreignKey: 'fileId' });
 
 module.exports = Chunk;
